chore: move components with directories into their own respective dirs
