+++ client/components/AppLayout/index.js
@@ -1,0 +1,5 @@
+export default from './AppLayout'
+
+export { default as PageWrapper } from './PageWrapper'
+
+export * from './AppLayout'
