+++ client/components/Modal/index.js
@@ -1,0 +1,4 @@
+export default from './Modal'
+
+export { default as ModalContent } from './ModalContent'
+export { default as ModalFooter } from './ModalFooter'
