+++ client/components/RatCard/RatCard.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types'
 
 
 // Component imports
-import { connect, actionStatus } from '../store'
+import { connect, actionStatus } from '../../store'
 import {
   selectRatById,
   selectShipsByRatId,
@@ -14,12 +14,12 @@ import {
   selectUserDisplayRatId,
   selectPageViewMetaById,
   withCurrentUserId,
-} from '../store/selectors'
-import { formatAsEliteDate } from '../helpers/formatTime'
-import classNames from '../helpers/classNames'
-import CardControls from './CardControls'
-import DefaultRatButton from './RatCard/DefaultRatButton'
-import InlineEditSpan from './InlineEditSpan'
+} from '../../store/selectors'
+import { formatAsEliteDate } from '../../helpers/formatTime'
+import classNames from '../../helpers/classNames'
+import CardControls from '../CardControls'
+import DefaultRatButton from './DefaultRatButton'
+import InlineEditSpan from '../InlineEditSpan'
