+++ client/pages/_error.js
@@ -7,7 +7,7 @@ import NextError from 'next/error'
 
 
 // Component imports
-import PageWrapper from '../components/PageWrapper'
+import PageWrapper from '../components/AppLayout/PageWrapper'
