+++ client/pages/acknowledgements.js
@@ -3,7 +3,7 @@ import React from 'react'
 
 
 // Component imports
-import PageWrapper from '../components/PageWrapper'
+import { PageWrapper } from '../components/AppLayout'
 import { dependencies } from '../../package.json'
