+++ client/pages/i-need-fuel.js
@@ -4,7 +4,7 @@ import React from 'react'
 
 // Component imports
 import { Link } from '../routes'
-import PageWrapper from '../components/PageWrapper'
+import { PageWrapper } from '../components/AppLayout'
 import Component from '../components/Component'
