+++ client/pages/index.js
@@ -7,8 +7,8 @@ import React from 'react'
 
 // Component imports
 import { Link } from '../routes'
+import { PageWrapper } from '../components/AppLayout'
 import { connect } from '../store'
-import PageWrapper from '../components/PageWrapper'
 import Carousel from '../components/Carousel'
