+++ client/pages/leaderboard.js
@@ -6,7 +6,7 @@ import React from 'react'
 
 
 // Component imports
-import PageWrapper from '../components/PageWrapper'
+import { PageWrapper } from '../components/AppLayout'
 import RatLeaderboardTable from '../components/RatLeaderboardTable'
