+++ client/pages/storefront/cart.js
@@ -14,7 +14,7 @@ import {
 } from '../../store/selectors'
 import { Link } from '../../routes'
 import Component from '../../components/Component'
-import PageWrapper from '../../components/PageWrapper'
+import { PageWrapper } from '../../components/AppLayout'
 import StoreControlBar from '../../components/storefront/StoreControlBar'
 import getMoney from '../../helpers/getMoney'
