+++ client/pages/storefront/checkout.js
@@ -8,11 +8,10 @@ import React from 'react'
 
 // Component imports
 import { actions, connect } from '../../store'
-import { withStripe } from '../../components/AppLayout'
+import { PageWrapper, withStripe } from '../../components/AppLayout'
 // import StoreControlBar from '../../components/storefront/StoreControlBar'
 import Component from '../../components/Component'
 // import CheckoutForm from '../../components/storefront/CheckoutForm'
-import PageWrapper from '../../components/PageWrapper'
 import { selectStoreCart } from '../../store/selectors'
