+++ client/pages/storefront/list.js
@@ -10,7 +10,7 @@ import { actions, connect } from '../../store'
 // import StoreControlBar from '../../components/storefront/StoreControlBar'
 import CartUpdateModal from '../../components/storefront/CartUpdateModal'
 import Component from '../../components/Component'
-import PageWrapper from '../../components/PageWrapper'
+import { PageWrapper } from '../../components/AppLayout'
 import ProductCard from '../../components/storefront/ProductCard'
 import productPriorityDecendingSort from '../../helpers/productPriorityDecendingSort'
 import { selectProductsMeta, selectProducts } from '../../store/selectors'
