+++ client/pages/wordpress-proxy.js
@@ -8,7 +8,7 @@ import React from 'react'
 // Component imports
 import { actions, connect, actionStatus } from '../store'
 import Component from '../components/Component'
-import PageWrapper from '../components/PageWrapper'
+import { PageWrapper } from '../components/AppLayout'
 import { selectWordpressPageBySlug } from '../store/selectors'
